backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfplumber
pytesseract
Pillow
pdf2image
anthropic
python-jose[cryptography]
bcrypt
passlib[bcrypt]
thefuzz
python-Levenshtein
python-dotenv
aiofiles
